Implemented DataStorageService with dispartion calculating functionality.

+++ src/main/java/com/example/mario3/entity/Point.java
@@ -4,12 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
 
 @Data
 @Builder
 @AllArgsConstructor
 public class Point {
-    BigDecimal x;
-    BigDecimal y;
+    Double x;
+    Double y;
 }
